Refactor for shared usage of the ErrorExitReplace structure
